moves inputs to atoms
