+++ src/components/atoms/inputs/Color/stories/colorInput.stories.tsx
